+++ src/Opulence/Sessions/Handlers/SessionEncryptionException.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2016 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+namespace Opulence\Sessions\Handlers;
+
+use Exception;
+
+/**
+ * Defines a session encryption exception
+ */
+class SessionEncryptionException extends Exception
+{
+    // Don't do anything
+}